Fix Github docs action for Google COS...
The Google COS manifest repo lists all versions including those not released, which causes an error in when it comes to downloading kernel headers and configs etc. This commit fixes that by validating the versions against the Google COS storage to ensure that the "latest" version passed through to the POC has released artifacts.