Fix COS kernel config location and build errors
Falco-driver-builder expects the kernel to be in one of several places of which we use `/lib/modules/<kernel_release>/config`. This commit fixes a bug where an extraneous + was added to the `/lib/modules/<kernel_release>` folder (resulting in `/lib/modules/<kernel_release>+/config`), meaning falco-driver-loader was unable to find the config.

In addition fixed several build errors mostly to newer kernel and older falco including:

* [`use of undeclared identifier 'KBUILD_MODNAME'`](https://github.com/falcosecurity/falco/issues/1334#issuecomment-718049958): Conditionally added KBUILD_MODNAME to `driver_config.h`.
* Kernel Makefile not calling bfp/Makefile ([`always` vs `alway-y`](https://github.com/falcosecurity/libs/blob/3.0.1+driver/driver/bpf/Makefile#L8-L10)): Conditionally added an `always-y` rule to bpf/Makefile
* `no member named 'audit' in 'struct task_struct'` ([caused by COS_73_WORKAROUND patch being misapplied to cos-101-17033-0-0 to cos-101-17109-0-0](https://github.com/draios/sysdig/pull/1431)): Removed builds for .0.0 COS versions as these are not officially released and likely alpha versions anyway.